Added color change when moving to next peson
When clicking the next button to choice another name from the hat, the background and foreground colors will change. This makes it easier on the eye to know a change has happend when the next button has been clicked.

Also made it so the next person picked can not be the current person that was picked. This would sometime happen when the current person on screen was the last name in the bag, and then when the bag of names got refiled it might pick the same person again.